AsyncComm: Moved timeout from Connect() to SendMessage()
Moved timeout parameter from Connect() to SendMessage() since
that is where it is needed and to allow requests to get sent
across connections that were established via Listen.  Also
modified sampleClient and testServer to have options for making
the server connect to the client to pave the way for a test
to verify that requests can get successfully sent over connections
that were established via Listen.


git-svn-id: https://hypertable.googlecode.com/svn/trunk@341 01261837-d133-0410-8dcb-6792eb736156