Got Server registration working
Implemented ServerRegistration command which is invoked by
RangeServers when they establish a connection to the master.
The argument's to this method are the range server's unique
server ID string and the IP address from which it connected.
This allows the Master to map the server's ID string to the
Comm layer connection address.


git-svn-id: https://hypertable.googlecode.com/svn/trunk@347 01261837-d133-0410-8dcb-6792eb736156