Overhauled MasterClient
Got MasterClient implemented such that it can be used by the Client Library
as well as the Range Server.  It now discovers the master location by reading
the /hypertable/master file's 'address' attribute.  It also now requires an
application queue so that when it receives an ATTR_SET event on the master
file, it can re-read that file without deadlocking by pushing the job onto the
app queue.


git-svn-id: https://hypertable.googlecode.com/svn/trunk@349 01261837-d133-0410-8dcb-6792eb736156