Added support for alternate commit log DFS broker;  Changed default ports
The RangeServer can be configured to write its commit log to a different
DFS broker.  This is primarily to overcome a limitation in Hadoop that
makes it unable to support a commit log.  The alternate dfs can be
selected with the --log-broker=<addr> command line option or the following
config file properties:
Hypertable.RangeServer.CommitLog.DfsBroker.host
Hypertable.RangeServer.CommitLog.DfsBroker.port
The default ports for all of the services have been changed to be
more memorable


git-svn-id: https://hypertable.googlecode.com/svn/trunk@369 01261837-d133-0410-8dcb-6792eb736156