Added configurable compression type to CREATE TABLE
Added a COMRESSOR="type args" argument to the ACCESS GROUP
clause.  Also added optional table options right after
CREATE TABLE with COMPRESSOR option.  The idea is that
compressors can be specified with arguments such as the
following:  COMPRESSOR="zlib --best".  Added all of the
plumbing to get it all to work.


git-svn-id: https://hypertable.googlecode.com/svn/trunk@508 01261837-d133-0410-8dcb-6792eb736156