Fixed lzo build on Linux change flags to type in compressor header
For whatever reason, the static lzo2 library is not being found on
Linux, so I temporarily just hard code it to -llzo2 in the toplevel
CMakeLists.txt file since we're going to just include the source
code soon.  Also got rid of the flags field in the BlockCompressionHeader
and replaced it with a type field.