Fixed problems with DELETE and SHOW TABLES
The first key in a CellCache or CellStore was getting
skipped if it was a ROW_DELETE and the scan was for
a specific column family.  There was also a small
problem that was causing show tables to not work
(passing a vector into a function by value, not by
reference).