Created new StaticBuffer class which is now used in all internal APIs
The StaticBuffer class is used to send buffers through API calls.
It has an 'own' flag that indicates whether or not the buffer should
get deleted in the destructor.  The class does explicit ownership management
and transfers ownership in the copy constructor and the assignment operator.
Also added an 'own' flag to DynamicBuffer.